tests/wpt/web-platform-tests/html/semantics/forms/the-input-element/search_input.html

<!DOCTYPE html>
<html>

  <head>
    <title>Search Input</title>
     <meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
     <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
     <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
    <link rel="help" href="https://html.spec.whatwg.org/multipage/#the-input-element">
   <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>


      <h1>Search Input</h1>
      <input type="search" style="display:none" placeholder="Search..." />

    <div id="log">
    </div>

  <script type="text/javascript">


    test(function() {assert_equals(document.getElementsByTagName("input")[0].type, "search")}, "search type support on input element", {
      "help" : "https://html.spec.whatwg.org/multipage/#dom-input-type" });
    test(function() {assert_equals(document.getElementsByTagName("input")[0].placeholder, "Search...")}, "placeholder attribute support on input element", {
      "help" : "https://html.spec.whatwg.org/multipage/#dom-input-placeholder" });

  </script>

  </body>

</html>
